+++ packages/cloudflare/src/request.ts
@@ -218,5 +218,5 @@ export function wrapRequestHandlerWithInit(
         });
       },
     );
-  });
+  }, wrapperOptions.context);
 }